Move test-specific methods to tests
